3619: Rename TransformEntityAttributesQuickFix files
