+++ common/src/Model/TransformEntityPropertiesQuickFix.h
